requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
deepagents>=0.4.0
langgraph>=0.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
langchain-tavily>=0.1.0
